Homework 6 fixes

+++ src/main/java/ru/mail/polis/homework/io/blocking/Structure.java
@@ -59,7 +59,7 @@ public class Structure {
         return coeff;
     }
 
-    public void setCoeff(float coeff) {
+    public void setCoeff(double coeff) {
         this.coeff = coeff;
     }
